The HUD renders
if/ is in the build and the in-match HUD draws: both players' percent
displays and stock icons, updating with damage and surviving a KO.

The SceneDesc converter was the entry fee; the rest was a chain of the
port's signature bug in different clothes.

HSD_SisLib_Alloc(0xA0) is sizeof(HSD_Text) on GameCube. The struct is
larger here, so every write past 0xA0 landed on the next block's SisBlock
header and corrupted sislib's free list -- the allocator then walked into
unmapped memory. Two more literal allocation sizes alongside it, plus
"- 0xC" for the pool header.

Two overlay structs in ifmagnify.c (pad[0x5C], pad[0x74]) reach
image_descs at hardcoded GameCube offsets. On x86_64 those land inside
player[], which is what wrote a host-heap address into player[3].gobj.

ifStatus_802F6194 walks an HSD_JObj* through HSD_GObj's fields. That is
only correct because GObj::next_gx and JObj::child are both at +0x10 on
GameCube, and ::next at +0x08 for both; HSD_Obj widens here and neither
holds. It returned garbage for every HUD digit joint.

Player_80036978(s32 slot, s32 arg1) takes a Vec3* in an s32 -- the decomp
marks it "@todo Eliminate cast" -- and truncated a stack address.

lbArchive_80016DBC dropped its variadic (symbol, name) pairs, so callers'
out-pointers kept whatever they held. OSPanic was declared noreturn while
its stub returns, the same bug __assert had, which is why sislib's
allocator recovery paths had been deleted. And pl_8003906C computes
"% arg3" with a period from an undecompiled bonus table that can be zero:
PowerPC's divw produced a garbage quotient, x86 traps.

ifMagnify_803F97E8 is a camera descriptor in .data satisfied by a weak
*function* stub, so &symbol was a code address read as a camera. An audit
finds 412 data symbols in that state; the roadmap records how to list
them. The magnifier stays disabled because ~20 of its float constants are
in that set -- ifMagnify_804DDB4C read back as 1.6e36 and handed
HSD_CObjSetOrtho a degenerate frustum.

Verified: 2200-frame match with the HUD up through damage, KO and
respawn, exit 0; characters 0/2/4/6/14/25 and stages 0/8/12/21/32 clean.

+++ src/sysdolphin/baselib/sislib.h
@@ -112,7 +112,9 @@ extern SisBlock* used_head;
 extern HSD_Text* HSD_SisLib_804D7978;
 extern sislib_UnkAlloc3* HSD_SisLib_804D797C;
 
-void HSD_SisLib_803A6048(size_t);
+/* u32, not size_t: the definition takes a 32-bit byte count, and the two
+ * agree only where size_t is 4 bytes. */
+void HSD_SisLib_803A6048(u32);
 UNK_T func_804A70A0(UNK_T, s32, u32, s32, s32, s32, s32, s32, f64, f64, f64,
                     f64, f64, f64, f64, f64);
 void HSD_SisLib_803A947C(HSD_Archive*);
